refactor(#3585): move all tests to one module
